+++ models/Card.js
@@ -1,3 +1,4 @@
+// deck.js
 const getCardModel = (sequelize, { DataTypes }) => {
   const Card = sequelize.define("card", {
     id: {
@@ -38,8 +39,27 @@ const getCardModel = (sequelize, { DataTypes }) => {
     },
   });
 
-  Card.associate = (models) => {
-    Card.belongsTo(models.Deck);
+  // Add a method to get a shuffled deck
+  // Add a method to get a shuffled deck
+  Card.getShuffledDeck = async () => {
+    try {
+      const cards = await Card.findAll(); // Get all cards from the database
+      if (!cards || cards.length === 0) {
+        throw new Error("No cards found in the database.");
+      }
+
+      // Shuffle the cards using the provided logic
+      for (let i = cards.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [cards[i], cards[j]] = [cards[j], cards[i]];
+      }
+
+      // console.log("Shuffled deck:", cards); // Log the shuffled deck for debugging
+      return cards;
+    } catch (error) {
+      console.error("Error getting shuffled deck:", error);
+      throw error;
+    }
   };
 
   return Card;
