+++ models/index.js
@@ -2,8 +2,8 @@ const sequelize = require("../config/connection");
 const Sequelize = require("sequelize");
 
 const { getUserModel } = require("./User");
-// const { getCardModel } = require("./Card");
-const { getDeckModel } = require("./Deck");
+const { getCardModel } = require("./Card");
+// const { getDeckModel } = require("./Deck");
 const { getTableModel } = require("./Table");
 const { getUserGroupRoleModel } = require("./UserGroupRole");
 const { getCashGameModel } = require("./CashGame");
@@ -23,8 +23,8 @@ const models = {
   CashGame: getCashGameModel(sequelize, Sequelize),
   TournamentGame: getTournamentGameModel(sequelize, Sequelize),
   PokerGroup: getPokerGroupModel(sequelize, Sequelize),
-  // Card: getCardModel(sequelize, Sequelize),
-  Deck: getDeckModel(sequelize, Sequelize),
+  Card: getCardModel(sequelize, Sequelize),
+  // Deck: getDeckModel(sequelize, Sequelize),
   Table: getTableModel(sequelize, Sequelize),
   BlindLevel: getBlindLevelModel(sequelize, Sequelize),
 };
