+++ seeders/seed.js
@@ -1,5 +1,5 @@
 const sequelize = require("../config/connection");
-const { User, UserGroupRole, PokerGroup, Deck } = require("../models");
+const { User, UserGroupRole, PokerGroup, Card } = require("../models");
 const fs = require("fs");
 const path = require("path");
 
@@ -47,7 +47,7 @@ const seedDatabase = async (sequelize) => {
     }
     // Seed Cards
     for (const cardData of cardSeedData) {
-      await Deck.create(cardData);
+      await Card.create(cardData);
     }
 
     console.log("Database seeded successfully.");
